pyproject.toml
[project]
name = "pinboard"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "pyside6>=6.10.1",
    "pyyaml>=6.0.3",
]
